% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_match_patients.R
\name{calc_match_patients}
\alias{calc_match_patients}
\title{Find all potential person matches between two datasets based on personal identifiable information}
\usage{
calc_match_patients(
  df_one,
  id_one,
  forename_one,
  surname_one,
  dob_one,
  postcode_one,
  df_two,
  id_two,
  forename_two,
  surname_two,
  dob_two,
  postcode_two,
  output_type = c("all", "key", "match"),
  format_data = c(TRUE, FALSE),
  inc_no_match = c(TRUE, FALSE),
  sw_forename = 0.3,
  sw_surname = 0.15,
  sw_dob = 0.4,
  sw_postcode = 0.15
)
}
\arguments{
\item{id_one}{unique reference column for df1}

\item{forename_one}{forename column for df1}

\item{surname_one}{surname column for df1}

\item{dob_one}{date of birth column for df1, ideally in dd/mm/yyyy format although others are
handled}

\item{postcode_one}{postcode field for df1}

\item{id_two}{unique reference column for df2}

\item{forename_two}{forename column for df2}

\item{surname_two}{surname column for df2}

\item{dob_two}{date of birth column for df2, ideally in dd/mm/yyyy format although others are
handled}

\item{postcode_two}{postcode field for df2}

\item{output_type}{One of the following: "key" / "match" / "all"}

\item{format_data}{TRUE/FALSE : identifying if the input datasets should be passed through
formatting functions to clean data prior to matching}

\item{inc_no_match}{TRUE/FALSE : identifying if the output should include non matches}

\item{sw_forename}{(default = 0.30) proportion weighting value (0.0-1.0) to be applied to the forename part of the match score}

\item{sw_surname}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the surname part of the match score}

\item{sw_dob}{(default = 0.40) proportion weighting value (0.0-1.0) to be applied to the dob part of the match score}

\item{sw_postcode}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the postcode part of the match score}

\item{df1}{dataframe containing person level information representing people to look for}

\item{df2}{dataframe containing person level information representing people to look in for
matches}
}
\value{
dataframe comprising of all potential matches between two datasets
}
\description{
The two input datasets should be dataframes, with the first dataset (df1) containing the people
to look for and the second dataset (df2) containing the data to be searched against. Parameters
can be used to specify the formatting of the output dataset and allow some customisation of the
match scoring.
}
\examples{
find_matches(df1, df2, ...)
}
