% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_match_person_self.R
\name{calc_match_person_self}
\alias{calc_match_person_self}
\title{Find all potential person matches within a single dataset based on personal identifiable information}
\usage{
calc_match_person_self(
  df,
  id,
  forename,
  surname,
  dob,
  postcode,
  output_type = c("all", "key", "match"),
  format_data = c(TRUE, FALSE),
  inc_no_match = c(TRUE, FALSE),
  unique_combinations_only = c(TRUE, FALSE),
  sw_forename = 0.3,
  sw_surname = 0.15,
  sw_dob = 0.4,
  sw_postcode = 0.15
)
}
\arguments{
\item{df}{dataframe containing person level information representing people to look for}

\item{id}{unique reference column for df}

\item{forename}{forename column for df}

\item{surname}{surname column for df}

\item{dob}{date of birth column for df, ideally in dd/mm/yyyy format although others are handled}

\item{postcode}{postcode field for df}

\item{output_type}{One of the following: "key" / "match" / "all"}

\item{format_data}{TRUE/FALSE : identifying if the input datasets should be passed through
formatting functions to clean data prior to matching}

\item{inc_no_match}{TRUE/FALSE : identifying if the output should include non matches}

\item{unique_combinations_only}{TRUE/FALSE : identifying if the output should include both directions for combination pairs (e.g. A=B & B=A)}

\item{sw_forename}{(default = 0.30) proportion weighting value (0.0-1.0) to be applied to the forename part of the match score}

\item{sw_surname}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the surname part of the match score}

\item{sw_dob}{(default = 0.40) proportion weighting value (0.0-1.0) to be applied to the dob part of the match score}

\item{sw_postcode}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the postcode part of the match score}
}
\value{
dataframe comprising of all potential matches within the dataset
}
\description{
The input dataset should be dataframe, where for each person in the dataset the function will
try to identify any other records for the same individual.
\cr\cr Match results will include cases where the records match either exactly or confidently based on matching rules.
\cr\cr Parameters can be used to specify the formatting of the output dataset and allow some
customisation of the match scoring.
}
