% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_match_person_self_db.R
\name{calc_match_person_self_db}
\alias{calc_match_person_self_db}
\title{Find all potential person matches within a single dataset based on personal identifiable information}
\usage{
calc_match_person_self_db(
  df,
  id,
  forename,
  surname,
  dob,
  postcode,
  output_type = c("all", "key", "match"),
  inc_no_match = c(TRUE, FALSE),
  unique_combinations_only = c(TRUE, FALSE),
  sw_forename = 0.3,
  sw_surname = 0.15,
  sw_dob = 0.4,
  sw_postcode = 0.15
)
}
\arguments{
\item{df}{a 'lazyframe' object containing person level information representing people to look for}

\item{id}{unique reference column within df}

\item{forename}{forename column within df}

\item{surname}{surname column within df}

\item{dob}{date of birth column within df, in format YYYYMMDD}

\item{postcode}{postcode field within df}

\item{output_type}{One of the following: "key" / "match" / "all"}

\item{inc_no_match}{TRUE/FALSE : identifying if the output should include non matches}

\item{unique_combinations_only}{TRUE/FALSE : identifying if the output should include both directions for combination pairs (e.g. A=B & B=A)}

\item{sw_forename}{(default = 0.30) proportion weighting value (0.0-1.0) to be applied to the forename part of the match score}

\item{sw_surname}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the surname part of the match score}

\item{sw_dob}{(default = 0.40) proportion weighting value (0.0-1.0) to be applied to the dob part of the match score}

\item{sw_postcode}{(default = 0.15) proportion weighting value (0.0-1.0) to be applied to the postcode part of the match score}
}
\value{
A 'lazyframe' comprising of all potential matches, to either be collected or written back to a database.
}
\description{
The input dataset should be a 'lazyframe', generated from a connection to a database. For each person in the dataset
the function will try to identify any other records for the same individual.
\cr\cr Parameters can be used to specify the formatting of the output dataset and allow some customisation of the match scoring.
\cr\cr Match results will include cases where the records match either exactly or confidently based on matching rules.
}
