% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_permutations.R
\name{calc_permutations}
\alias{calc_permutations}
\title{Calculates 9 permutations for primary-lookup join prior to match scoring}
\usage{
calc_permutations(df, forename, surname, postcode, dob)
}
\arguments{
\item{df}{A df to be formatted}

\item{forename}{forename column name}

\item{surname}{surname column name}

\item{postcode}{postcode column name}

\item{dob}{DOB column name}
}
\value{
A df with 5 'basic' join permutations added
}
\description{
Calculates 9 permutations for primary-lookup join prior to match scoring
}
\examples{
calc_permutations(df, forename, surname, postcode, dob)
}
