% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dob.R
\name{filter_dob}
\alias{filter_dob}
\title{Applies all combinations of dob_substr() filters}
\usage{
filter_dob(df, dob_one, dob_two, diff_threshold)
}
\arguments{
\item{df}{A df with dates to be formatted}
}
\value{
A df with relevant dates from both columns
}
\description{
Through doing so, it mirrors an edit distance of 2 on an 8-digit string
For this very particular use case, is faster than LV within SQL
}
\examples{
dob_filter(df, dob_one, dob_two, dob_diff)
}
