% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dob_db.R
\name{filter_dob_db}
\alias{filter_dob_db}
\title{Applies all combinations of dob_substr() filters}
\usage{
filter_dob_db(df, dob_one, dob_two, dob_score_threshold)
}
\arguments{
\item{df}{A df with dates to be formatted}
}
\value{
A df with relevant dates from both columns
}
\description{
Through doing so, it mirrors an edit distance of 2 on an 8-digit string
For this very particular use case, is faster than LV within SQL
}
\examples{
filter_dob_db(df, dob_one, dob_two, dob_score_threshold)
}
