% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_name_db.R
\name{filter_name_db}
\alias{filter_name_db}
\title{Applies a filter on either forename or surname to limit cross-join}
\usage{
filter_name_db(df, name_one, name_two)
}
\arguments{
\item{df}{A df to be formatted}

\item{name_one}{first name column}

\item{name_two}{second name column}
}
\value{
A df with a filtered name-col to limit post cross-join
}
\description{
Only contains name-pair instances that share certain characteristsics
These include same 1st, 2nd or last letter, or being a substring of another
}
\examples{
name_db_filter(df, name_one, name_two)
}
