% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matches.R
\name{find_matches}
\alias{find_matches}
\title{Find all potential matches between two datasets}
\usage{
find_matches(df1, df2)
}
\arguments{
\item{df1}{A dataframe containing person level information}

\item{df2}{A dataframe containing person level information}
}
\value{
dataframe comprising of all potential matches between two datasets
}
\description{
Find all potential matches between two datasets
}
\examples{
find_matches(df1, df2)
}
