% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMatches.R
\name{FindMatches}
\alias{FindMatches}
\title{FindMatches
Find all potential matches between two datasets}
\usage{
FindMatches(df1, df2)
}
\arguments{
\item{df1}{A dataframe containing person level information}

\item{df2}{A dataframe containing person level information}
}
\value{
dataframe comprising of all potential matches between the two datasets
}
\description{
FindMatches
Find all potential matches between two datasets
}
\examples{
FindMatches(df1, df2)
}
