% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_homoglyph.R
\name{fix_homoglyph}
\alias{fix_homoglyph}
\title{Fix common homoglyph errors that appear}
\usage{
fix_homoglyph(string, char_pos, character_or_number)
}
\arguments{
\item{string}{A string to be corrected}

\item{char_pos}{The character index position to be corrected}

\item{character_or_number}{Whether character or numbers should be corrected}
}
\value{
A character where known homoglyphs have been switched
}
\description{
Fix common homoglyph errors that appear
}
\examples{
fix_homoglyph(string, char_pos, character_or_number)
}
