% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixHomoglyph.R
\name{FixHomoglyph}
\alias{FixHomoglyph}
\title{FixHomoglyph
Fix common homoglyph errors that appear}
\usage{
FixHomoglyph(x)
}
\arguments{
\item{x}{A character to be fixed}
}
\value{
A character where known homoglyphs have been switched
}
\description{
FixHomoglyph
Fix common homoglyph errors that appear
}
\examples{
FixHomoglyph(x)
}
