% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Format the date of birth prior to matching}
\usage{
format_date(df, date)
}
\arguments{
\item{df}{a dataframe to feed into function}

\item{dob}{A vector of type date, character or integer with date expressions}
}
\value{
A dataframe with cleansed date (converted to YYYYMMDD) information
}
\description{
Formatting converts date field to string using format YYYYMMDD
\cr\cr Can handle dates supplied in a range of formats
}
