% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date_db.R
\name{format_date_db}
\alias{format_date_db}
\title{Format a DOB within the DB}
\usage{
format_date_db(df, date_col)
}
\arguments{
\item{df}{A df to be formatted}

\item{name_col}{a date column}
}
\value{
A df with cleansed date information, as a 8-digit number
}
\description{
Format the date of birth prior to matching
Convert to a string representation in the format YYYYMMDD
}
\examples{
format_db_date(df, date_col)
}
