% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date_parse_db.R
\name{format_date_parse_db}
\alias{format_date_parse_db}
\title{Format a DOB within the DB: Parsing collected data then pushed back}
\usage{
format_date_parse_db(df, date)
}
\arguments{
\item{df}{A df to be formatted}

\item{name_col}{a date column}
}
\value{
A df with cleansed date information, as a 8-digit number
}
\description{
Format the date of birth prior to matching
Convert to a string representation in the format YYYYMMDD
This function will not work with sizeable database tables
This is due to the having to copy the table in memory
}
\examples{
format_date_parse_db(df, date_col)
}
