% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dob.R
\name{format_dob}
\alias{format_dob}
\title{Convert date to character yyyymmdd format.}
\usage{
format_dob(dob)
}
\arguments{
\item{dob}{A vector of type character, integer or numeric with date expressions}
}
\value{
A formatted date (yyyymmdd)
}
\description{
Convert date to character yyyymmdd format.
}
\examples{
format_dob(dob = "2019-Mar-21")
}
