% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_name.R
\name{format_name}
\alias{format_name}
\title{Format name strings prior to matching}
\usage{
format_name(df, name)
}
\arguments{
\item{df}{a dataframe to feed into function}

\item{name}{a string field to be cleansed}
}
\value{
A dataframe with cleansed name string
}
\description{
Suitable for formatting forename/surname strings prior to matching
\cr\cr Formatting includes conversion to upper case and removal of non-alphabetic characters
}
