% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_name_db.R
\name{format_name_db}
\alias{format_name_db}
\title{Format name strings prior to matching}
\usage{
format_name_db(df, name_col)
}
\arguments{
\item{df}{a 'lazyframe' generated from a database connection with postcode to be formatted}

\item{name_col}{a string field to be cleansed}
}
\value{
a 'lazyframe' object with cleansed name string data
}
\description{
Suitable for formatting forename/surname strings prior to matching
\cr\cr Formatting includes conversion to upper case and removal of non-alphabetic characters
}
