% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode_db.R
\name{format_postcode_db}
\alias{format_postcode_db}
\title{Format the postcode strings prior to matching}
\usage{
format_postcode_db(df, postcode)
}
\arguments{
\item{df}{A 'lazyframe' generated from a database connection with postcode to be formatted}

\item{postcode}{the postcode column to be formatted}
}
\value{
A cleansed string
}
\description{
Formatting includes conversion to upper case and removal of
non alphanumeric characters
}
\examples{
format_postcode_db(df, postcode)
}
