% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDOB.R
\name{FormatDOB}
\alias{FormatDOB}
\title{FormatDOB
Format the date of birth prior to matching
Convert to a string representation in the format YYYYMMDD}
\usage{
FormatDOB(dob)
}
\arguments{
\item{dob}{A field to be cleansed}
}
\value{
A cleansed string
}
\description{
FormatDOB
Format the date of birth prior to matching
Convert to a string representation in the format YYYYMMDD
}
\examples{
FormatDOB(dob)
}
