% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatName.R
\name{FormatName}
\alias{FormatName}
\title{FormatName
Format the name strings prior to matching
Formatting includes conversion to upper case and removal of non alphabetic characters}
\usage{
FormatName(strName)
}
\arguments{
\item{strName}{A string field to be cleansed}
}
\value{
A cleansed string
}
\description{
FormatName
Format the name strings prior to matching
Formatting includes conversion to upper case and removal of non alphabetic characters
}
\examples{
FormatName(strName)
}
