% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPostcode.R
\name{FormatPostcode}
\alias{FormatPostcode}
\title{FormatPostcode
Format the postcode strings prior to matching
Formatting includes conversion to upper case and removal of non alphanumeric characters}
\usage{
FormatPostcode(strPostcode)
}
\arguments{
\item{strPostcode}{A string field to be cleansed}
}
\value{
A cleansed string
}
\description{
FormatPostcode
Format the postcode strings prior to matching
Formatting includes conversion to upper case and removal of non alphanumeric characters
}
\examples{
FormatName(strPostcode)
}
