% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TEST_DF_B}
\alias{TEST_DF_B}
\title{TEST_DF_B}
\format{
A data frame with 4 rows and 5 variables:
\describe{
  \item{\code{ID}}{character Unique reference column}
  \item{\code{SURNAME}}{character Surname of individual}
  \item{\code{FORENAME}}{character Forename of individual}
  \item{\code{POSTCODE}}{character Residential address postcode}
  \item{\code{DOB}}{character Date of birth}
}
}
\usage{
TEST_DF_B
}
\description{
Example dummy data for persons to be reviewed for potential matches
}
\keyword{datasets}
