% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_permutations.R
\name{calc_permutations}
\alias{calc_permutations}
\title{Create permutations of personal data for matching}
\usage{
calc_permutations(df, forename, surname, postcode, dob)
}
\arguments{
\item{df}{df to be formatted}

\item{forename}{forename column name}

\item{surname}{surname column name}

\item{postcode}{postcode column name}

\item{dob}{DOB column name}
}
\value{
df with join permutations added
}
\description{
This is a support function called during the execution of the calc_match_patients function.
\cr\cr Calculates permutations of personal information to support joining datasets, allowing
data to be initially joined only where the permutations of personal information align.
\cr\cr The permutations range from combinations of full fields to only partial information
extracted from each piece of personal information.
}
