% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Convert date to character yyyymmdd format.}
\usage{
format_date(df, date)
}
\arguments{
\item{dob}{A vector of type character, integer or numeric with date expressions}
}
\value{
A formatted date (yyyymmdd)
}
\description{
Convert date to character yyyymmdd format.
}
\examples{
format_dob(df, dob)
}
