% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date_db.R
\name{format_date_db}
\alias{format_date_db}
\title{Format the date of birth prior to matching}
\usage{
format_date_db(df, date_col)
}
\arguments{
\item{df}{a 'lazyframe' generated from a database connection with date field to be formatted}

\item{date_col}{a date column to be formatted}
}
\value{
a 'lazyframe' object with cleansed date of birth field
}
\description{
Formatting converts date field to numeric format using format YYYYMMDD
\cr\cr Date should be based on a recognised date type
}
