% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date_parse_db.R
\name{format_date_parse_db}
\alias{format_date_parse_db}
\title{Format a DOB within the DB: Parsing collected data then pushed back}
\usage{
format_date_parse_db(df, date)
}
\arguments{
\item{df}{a 'lazyframe' generated from a database connection with date field to be formatted}

\item{date}{a date column to be formatted}
}
\value{
a 'lazyframe' object with cleansed date of birth field
}
\description{
Formatting converts date field to string using format YYYYMMDD
\cr\cr Can handle dates supplied in a range of formats
\cr\cr This function will not work with sizeable database tables due to having to copy the table
in memory
}
