% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dob.R
\name{format_dob}
\alias{format_dob}
\title{Format a DOB}
\usage{
format_dob(dob)
}
\arguments{
\item{dob}{A date of birth date field to be cleansed}
}
\value{
A cleansed string
}
\description{
Format the date of birth prior to matching
Convert to a string representation in the format YYYYMMDD
}
\examples{
format_dob(dob)
}
