% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_name.R
\name{format_name}
\alias{format_name}
\title{Formatting either a Forename or Surname}
\usage{
format_name(str_name)
}
\arguments{
\item{str_name}{A string field to be cleansed}
}
\value{
A cleansed string
}
\description{
Format the name strings prior to matching
Formatting includes conversion to upper case and removal of
non alphabetic characters
}
\examples{
format_name(str_name)
}
