% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_name_db.R
\name{format_name_db}
\alias{format_name_db}
\title{Formatting either a Forename or Surname within the DB}
\usage{
format_name_db(df, name_col)
}
\arguments{
\item{df}{A df to be formatted}

\item{name_col}{a patient name column}
}
\value{
A df with cleansed patient name information
}
\description{
Format the name strings prior to matching
Formatting includes conversion to upper case and removal of
non alphabetic characters
}
\examples{
format_db_name(df, name_col)
}
