% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode.R
\name{format_postcode}
\alias{format_postcode}
\title{Format the postcode strings prior to matching}
\usage{
format_postcode(df, id, postcode)
}
\arguments{
\item{df}{A dataframe to feed into function}
}
\value{
A dataframe with cleansed postcode information
}
\description{
Formatting includes conversion to upper case and removal of
non alphanumeric characters
}
\examples{
format_postcode(df)
}
