% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode.R
\name{format_postcode}
\alias{format_postcode}
\title{Format the postcode strings prior to matching}
\usage{
format_postcode(postcode)
}
\arguments{
\item{strPostcode}{A string field to be cleansed}
}
\value{
A cleansed string
}
\description{
Formatting includes conversion to upper case and removal of
non alphanumeric characters
}
\examples{
format_postcode(postcode)
}
